Remove UnifiedPlugin
The UnifiedPlugin was an attempt to have each testbed apply the same plugin.
However, with this approach, we don't get generated accessors.
We will want to follow this approach in general, however this will be implemented by a new feature natively in gradle
