+++ unified-prototype/testbed-android-groovy/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id("org.gradle.unified-prototype")
+    id("org.gradle.android-prototype")
 }
 
 androidLibrary {
