+++ unified-prototype/testbed-jvm-groovy/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id("org.gradle.unified-prototype")
+    id("org.gradle.jvm-prototype")
 }
 
 jvmLibrary {
