+++ unified-prototype/unified-plugin/plugin-jvm/src/main/java/org/gradle/api/experimental/jvm/StandaloneJvmPlugin.java
@@ -1,0 +1,18 @@
+package org.gradle.api.experimental.jvm;
+
+import org.gradle.api.Plugin;
+import org.gradle.api.Project;
+
+/**
+ * Creates a declarative {@link JvmLibrary} DSL model, applies the official Jvm plugin,
+ * and links the declarative model to the official plugin.
+ */
+public class StandaloneJvmPlugin implements Plugin<Project> {
+    @Override
+    public void apply(Project project) {
+        project.getExtensions().create("jvmLibrary", JvmLibrary.class);
+        System.out.println("Applied JVM!");
+
+        // TODO: Actually do something here
+    }
+}
