--- unified-prototype/unified-plugin/plugin-jvm/src/main/kotlin/org/gradle/api/experimental/jvm/StandaloneJvmPluginExtensions.kt
@@ -1,11 +1,0 @@
-package org.gradle.api.experimental.jvm
-
-import org.gradle.api.Project
-import org.gradle.api.experimental.jvm.internal.JvmLibraryAccessor
-
-/**
- * Used to access a declarative JVM library from the Kotlin DSL
- */
-fun Project.jvmLibrary(configure: JvmLibrary.() -> Unit) {
-    extensions.getByType(JvmLibraryAccessor::class.java).access().configure()
-}
