+++ unified-prototype/unified-plugin/plugin-kmp/build.gradle.kts
@@ -7,3 +7,12 @@ description = "Implements the declarative KMP DSL prototype"
 dependencies {
     implementation(project(":plugin-common"))
 }
+
+gradlePlugin {
+    plugins {
+        create("kmp-plugin") {
+            id = "org.gradle.kmp-prototype"
+            implementationClass = "org.gradle.api.experimental.kmp.StandaloneKmpPlugin"
+        }
+    }
+}
