+++ unified-prototype/unified-plugin/settings.gradle.kts
@@ -11,4 +11,3 @@ include("plugin-jvm")
 include("plugin-kmp")
 
 include("plugin-common")
-include("plugin-public")
